remove actions_seth, move them to actions
