+++ integration-tests/smoke/automation_test.go
@@ -25,7 +25,6 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/networks"
 	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
 
-	"github.com/smartcontractkit/chainlink/integration-tests/actions"
 	"github.com/smartcontractkit/chainlink/integration-tests/actions/automationv2"
 	"github.com/smartcontractkit/chainlink/integration-tests/contracts"
 	"github.com/smartcontractkit/chainlink/integration-tests/contracts/ethereum"
@@ -37,7 +36,7 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ocr2keeper/evmregistry/v21/gasprice"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ocr2keeper/evmregistry/v21/mercury/streams"
 
-	actions_seth "github.com/smartcontractkit/chainlink/integration-tests/actions/seth"
+	"github.com/smartcontractkit/chainlink/integration-tests/actions"
 )
 
 const (
@@ -46,6 +45,20 @@ const (
 	automationExpectedData          = "abcdef"
 	defaultAmountOfUpkeeps          = 2
 )
+
+// UpgradeChainlinkNodeVersions upgrades all Chainlink nodes to a new version, and then runs the test environment
+// to apply the upgrades
+func upgradeChainlinkNodeVersionsLocal(
+	newImage, newVersion string,
+	nodes ...*test_env.ClNode,
+) error {
+	for _, node := range nodes {
+		if err := node.UpgradeVersion(newImage, newVersion); err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
 func automationDefaultRegistryConfig(c tc.AutomationTestConfig) contracts.KeeperRegistrySettings {
 	registrySettings := c.GetAutomationConfig().AutomationConfig.RegistrySettings
@@ -127,7 +140,7 @@ func SetupAutomationBasic(t *testing.T, nodeUpgrade bool) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -171,7 +184,7 @@ func SetupAutomationBasic(t *testing.T, nodeUpgrade bool) {
 			startTime := time.Now()
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			// TODO Tune this timeout window after stress testing
@@ -195,7 +208,7 @@ func SetupAutomationBasic(t *testing.T, nodeUpgrade bool) {
 				expect := 5
 				// Upgrade the nodes one at a time and check that the upkeeps are still being performed
 				for i := 0; i < 5; i++ {
-					err = actions.UpgradeChainlinkNodeVersionsLocal(*cfg.GetChainlinkUpgradeImageConfig().Image, *cfg.GetChainlinkUpgradeImageConfig().Version, a.DockerEnv.ClCluster.Nodes[i])
+					err = upgradeChainlinkNodeVersionsLocal(*cfg.GetChainlinkUpgradeImageConfig().Image, *cfg.GetChainlinkUpgradeImageConfig().Version, a.DockerEnv.ClCluster.Nodes[i])
 					require.NoError(t, err, "Error when upgrading node %d", i)
 					time.Sleep(time.Second * 10)
 					expect = expect + 5
@@ -266,7 +279,7 @@ func TestSetUpkeepTriggerConfig(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -288,7 +301,7 @@ func TestSetUpkeepTriggerConfig(t *testing.T) {
 			}
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			l.Info().Msg("Waiting for all upkeeps to perform")
@@ -447,7 +460,7 @@ func TestAutomationAddFunds(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -461,7 +474,7 @@ func TestAutomationAddFunds(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			l.Info().Msg("Making sure for 2m no upkeeps are performed")
@@ -525,7 +538,7 @@ func TestAutomationPauseUnPause(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -539,7 +552,7 @@ func TestAutomationPauseUnPause(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -624,7 +637,7 @@ func TestAutomationRegisterUpkeep(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -638,7 +651,7 @@ func TestAutomationRegisterUpkeep(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			var initialCounters = make([]*big.Int, len(upkeepIDs))
@@ -659,7 +672,7 @@ func TestAutomationRegisterUpkeep(t *testing.T) {
 				}
 			}, "4m", "1s").Should(gomega.Succeed()) // ~1m for cluster setup, ~1m for performing each upkeep once, ~2m buffer
 
-			newConsumers, _ := actions_seth.RegisterNewUpkeeps(t, a.ChainClient, a.LinkToken,
+			newConsumers, _ := actions.RegisterNewUpkeeps(t, a.ChainClient, a.LinkToken,
 				a.Registry, a.Registrar, automationDefaultUpkeepGasLimit, 1)
 
 			// We know that newConsumers has size 1, so we can just use the newly registered upkeep.
@@ -718,7 +731,7 @@ func TestAutomationPauseRegistry(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -732,7 +745,7 @@ func TestAutomationPauseRegistry(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -796,7 +809,7 @@ func TestAutomationKeeperNodesDown(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -810,7 +823,7 @@ func TestAutomationKeeperNodesDown(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -902,7 +915,7 @@ func TestAutomationPerformSimulation(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumersPerformance, _ := actions_seth.DeployPerformanceConsumers(
+			consumersPerformance, _ := actions.DeployPerformanceConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -918,7 +931,7 @@ func TestAutomationPerformSimulation(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -973,7 +986,7 @@ func TestAutomationCheckPerformGasLimit(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumersPerformance, upkeepIDs := actions_seth.DeployPerformanceConsumers(
+			consumersPerformance, upkeepIDs := actions.DeployPerformanceConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -989,7 +1002,7 @@ func TestAutomationCheckPerformGasLimit(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -1128,7 +1141,7 @@ func TestUpdateCheckData(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			performDataChecker, upkeepIDs := actions_seth.DeployPerformDataCheckerConsumers(
+			performDataChecker, upkeepIDs := actions.DeployPerformDataCheckerConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -1141,7 +1154,7 @@ func TestUpdateCheckData(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
@@ -1209,7 +1222,7 @@ func TestSetOffchainConfigWithMaxGasPrice(t *testing.T) {
 			sb, err := a.ChainClient.Client.BlockNumber(context.Background())
 			require.NoError(t, err, "Failed to get start block")
 
-			consumers, upkeepIDs := actions_seth.DeployConsumers(
+			consumers, upkeepIDs := actions.DeployConsumers(
 				t,
 				a.ChainClient,
 				a.Registry,
@@ -1223,7 +1236,7 @@ func TestSetOffchainConfigWithMaxGasPrice(t *testing.T) {
 			)
 
 			t.Cleanup(func() {
-				actions_seth.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
+				actions.GetStalenessReportCleanupFn(t, a.Logger, a.ChainClient, sb, a.Registry, registryVersion)()
 			})
 
 			gom := gomega.NewGomegaWithT(t)
